start-src-dirtree - factor main loop out of application code
TODAY WE RECLAIM THE APPLICATION
TOMORROW OBJECT ORIENTATION
