+++ src/stulto-application.h
@@ -1,0 +1,14 @@
+//
+// Created by marcxjo on 9/28/22.
+//
+
+#ifndef APPLICATION_H
+#define APPLICATION_H
+
+#include <glib.h>
+
+gboolean stulto_application_init(int argc, char *argv[]);
+
+int stulto_get_exit_status();
+
+#endif //APPLICATION_H
